+++ appinfo/application.php
@@ -78,7 +78,8 @@ class Application extends App
                 $c->query("AppName"),
                 $c->query("Request"),
                 $c->query("UserId"),
-                $c->query("EpisodeMapper")
+                $c->query("EpisodeMapper"),
+                $c->query("FeedMapper")
             );
         });