+++ db/episodemapper.php
@@ -22,6 +22,7 @@
 
 namespace OCA\Podcasts\Db;
 
+use OCP\AppFramework\Db\DoesNotExistException;
 use OCP\IDBConnection;
 use OCP\AppFramework\Db\Mapper;
 
@@ -36,11 +37,11 @@ class EpisodeMapper extends Mapper
     /**
      * Constructor
      *
-     * @param IDBConnection $db
+     * @param IDBConnection $dbConnection
      */
-    public function __construct(IDBConnection $db)
+    public function __construct(IDBConnection $dbConnection)
     {
-        parent::__construct($db, "podcasts_episodes");
+        parent::__construct($dbConnection, "podcasts_episodes");
     }
 
     /**
@@ -78,22 +79,21 @@ class EpisodeMapper extends Mapper
     }
 
     /**
-     * Gets a list of episodes for the current user (limited by feed if ID is supplied)
+     * Gets a list of episodes for the current user (limited by feed if ID is
+     * supplied)
      *
      * @param string $uid
      * @param int    $feedId
-     * @param int    $limit
-     * @param int    $offset
      *
      * @return array
      */
-    public function getEpisodes($uid, $feedId = null, $limit = null, $offset = null)
+    public function getEpisodes($uid, $feedId = null)
     {
         $params = [$uid];
 
-        $sql = "SELECT *PREFIX*podcasts_episodes.*, *PREFIX*podcasts_feeds.cover 
-                FROM *PREFIX*podcasts_episodes 
-                INNER JOIN *PREFIX*podcasts_feeds ON *PREFIX*podcasts_episodes.feed_id = *PREFIX*podcasts_feeds.id 
+        $sql = "SELECT *PREFIX*podcasts_episodes.*, *PREFIX*podcasts_feeds.cover
+                FROM *PREFIX*podcasts_episodes
+                INNER JOIN *PREFIX*podcasts_feeds ON *PREFIX*podcasts_episodes.feed_id = *PREFIX*podcasts_feeds.id
                 WHERE *PREFIX*podcasts_episodes.uid = ?";
 
         if (false === is_null($feedId)) {
@@ -114,21 +114,41 @@ class EpisodeMapper extends Mapper
      * Updates the playback position of an episode
      *
      * @param string $uid
-     * @param int    $id
+     * @param int    $episodeId
      * @param int    $second
      * @param int    $duration
      *
      * @return \PDOStatement
      */
-    public function updatePosition($uid, $id, $second, $duration)
+    public function updatePosition($uid, $episodeId, $second, $duration)
     {
-        $id = (int)$id;
+        $episodeId = (int)$episodeId;
         $second = (int)$second;
         $duration = (int)$duration;
 
         return $this->execute(
             "UPDATE *PREFIX*podcasts_episodes SET current_second = ?, duration = ? WHERE id = ? AND uid = ?",
-            [$second, $duration, $id, $uid]
+            [$second, $duration, $episodeId, $uid]
+        );
+    }
+
+    /**
+     * Mark episode as played / revoke played flag
+     *
+     * @param string $uid
+     * @param int    $episodeId
+     * @param bool   $status
+     *
+     * @return \PDOStatement
+     */
+    public function updatePlayedStatus($uid, $episodeId, $status)
+    {
+        $episodeId = (int)$episodeId;
+        $status = (bool)$status;
+
+        return $this->execute(
+            "UPDATE *PREFIX*podcasts_episodes SET played = ? WHERE id = ? AND uid = ?",
+            [$status, $episodeId, $uid]
         );
     }
 
@@ -153,17 +173,27 @@ class EpisodeMapper extends Mapper
     /**
      * Loads an episode by it's ID
      *
-     * @param int    $id
+     * @param int    $episodeId
      * @param string $uid
      *
-     * @return \OCP\AppFramework\Db\Entity
+     * @throws DoesNotExistException
+     *
+     * @return array
      */
-    public function getEpisode($id, $uid)
+    public function getEpisode($episodeId, $uid)
     {
-        $id = (int)$id;
+        $episodeId = (int)$episodeId;
 
         $sql = "SELECT * FROM *PREFIX*podcasts_episodes WHERE id = ? AND uid = ?";
+        $stmt = $this->execute($sql, [$episodeId, $uid]);
 
-        return $this->findEntity($sql, [$id, $uid]);
+        $episode = $stmt->fetch();
+
+        if (false === $episode) {
+            throw new DoesNotExistException("Episode id={$episodeId}
+            uid={$uid} not found");
+        }
+
+        return $episode;
     }
 }