+++ CHANGELOG.md
@@ -1,7 +1,22 @@
+0.8.2
+
+* Various Bugfixes
+* Migrated to SCSS
+* Migrated to AngularJS
+* Migrated to CoffeeScript
+* Integrated CodeClimate
+* Integrated Bower
+* Integrated Grunt
+* Using Videogular as audio/video player
+* Episodes are now marked as played when reaching the last minute of playback
+* Added placeholder cover
+
 0.8.1
+
 * PHPUnit coverage
 * SQLite compatibility
 * pgsql compatibility
 
 0.8.0
+
 * First release